Split out a ResetNextOpMode method from OSSelect (which now calls it).
This allows other things that don't need to change the OS, but *do* need
to make sure that a mode transition gets setup if necessary, to setup that
transition too.  Currently the only case that needs this is push-based
reloads, when the reload has been setup (and then not performed because it
was optimized away).