Rearranging deck chairs.
Get rid of a noxious warning from the boot config when rc.healthd fails
because it is not supported on...well, just about anything.

Also noticed that the long dead nodetype script was about 10 years out of
date with respect to what nodes it recognized, so for some strange reason
I was compelled to fix it.

Leigh is about to get rid of these anyway, but I thought I would at least
screw up his merge!