Better handling of media and speeds with ethtool for linux.
Allow tmcd to specify speeds of 0 with a warning, and then use
autonegotiation to select media and speed.  Also, if the speed was Gb,
turn on autonegotiation first, then force speed/duplex settings.  This
has to be done because you cannot run at Gb without enabling
autonegotiation -- it's part of the protocol.