A couple of fixes for linktest on openvz vnodes.
Don't try to run rude/crude with increased priority as it will fail.
For truely unshaped links (tb-set-noshaping + 0 delay + 0 loss) don't
check for a specific latency as it can be extremely low. We still do the
ping tests, we just check to make sure that ping packets get through without
checking the RTT; i.e., it is not just a connectivity test.