Tweaks to default layer2 HP switch config.
Set snmp community appropriately.  Turn on spanning tree.

Also make sure that VLAN 1 hasn't gotten a control net addr via DHCP.
When we remove the current switch config and the switch reverts to
its default, it turns on DHCP for VLAN 1.  Then, if it manages to get
a control net address (i.e., from the new node pool), we have to make
sure to unset any IP address in VLAN 1 before assigning one in VLAN 999
(the control net VLAN we use in this static configuration).  Otherwise we
might end up trying to configure an IP addr in the same subnet on two
different VLANs, which the switch won't allow.