Add machinery to automatically setup mrouted on the control network.
This is specifically intended for elabinelab right now where configurations
with an "inner control network" need a multicast router for inner frisbee
to work. There is (an undocumented) NEEDMROUTED defs variable that controls
whether the machinery is triggered or not. It is not documented because, though
in theory this would work for a real boss setup, it has not been tested and
each install would probably need a custom mrouted.conf file.

For an inner elab, this variable is automatically set in defs-elabinelab
if there is an inner control network. The result is that the mrouted port
will be installed, it will be enabled in rc.conf, and a stub mrouted.conf
file is created (to force mrouted to ignore the real control network).