Bug fixes and per-experiment VLAN removal.
Now that we have logical wires, we *must* actually make sure that the
VLANs created on the per-experiment switch are removed at swapout (i.e.,
are at least removed from the vlans table), so that any logical wires can
be safely removed.  For now, we actually invoke snmpit to remove them.
This is probably a bad idea, because who knows what state the user left
the switch in prior to swapout.

The bugfixes are just me forgetting that \ does not continue a line in
perl :).