MLE: set port speeds for non user-specified segment/path endpoint interfaces.
We now make a pass through the implemented_by tree setting port speeds
that the user left up to assign (i.e., that are 0).  We start at the
lowest level links and assign default speeds.  Then the speed values for
the start and end segments for each path get propagated to the link
being implemented by the path.  Should be pretty generic for any number
of layers.  We warn if the higher level links don't support those speed
settings (if they don't, we're gonna have a problem; but I think assign
protects this from happening).