Install same bins/libs on clients as on control nodes.
One difference: we don't make all the symlinks to script_wrapper.py (like
node_avail et al).

Also, this is going to require m2crypto -- the right version -- on client
images.  If the scripts don't work on many images for now because
m2crypto is not there, fine for awhile... but we should move in this
direction.