Restart dhcpd on all subbosses
When rebuilding the dhcpd.conf file, make sure to restart dhcpd on all
subbosses and not just the ones that are listed in the subbosses
table.  Otherwise if all nodes are moved from one subboss to another
the old subboss will continue to try serving dhcp leases when it
shouldn't.