Added compat option and assorted cleanups.
Added WITH_V3COMPAT to make sure that we can still generated V3 images
(for other sites) if checksum/encryption is not used.

Try to clean up the command line options some. Be more consistent by putting
generated uuid into a file instead of just spitting it out on stderr. Make
sure that if the decryption specifies signing and/or encryption that we
require the image to have that info. Add some more assertions. (Re)distinguish
checksums from signed-checksums.