Split max_sliver_lifetime sitevar into max_sliver_lifetime (for the CM) and max_slice_lifetime (for the SA) so that slice credentials and slivers can have max different expirations. This is so a CM can can limit easily limit slivers so they do not hang out for months doing nothing.
Add call to ipassign_wrapper to fill in IPs since we are not running
the parser, and the stuff that is in the CM is bogus. Actually need to
remove that. Also need to do this for new bridge support.

Check new users for a valid email, so that we can send warnings from
the expire daemon when slivers are going to expires (see above).
Anyone from an Emulab will have an email address in their certificate,
but PLC users do not. Try to contact the PLC registry to resolve the
user, and fill in the email. If we cannot get the email, we fail.
This is not tested yet, cause PLC does not trust the CM certificate,
and this is probably the case for PLE and MAX, but this will force
things to get fixed.