Add an 'archivemail' configuration option
'archivemail' works exactly like 'alwaysmail', except that any
addresses it contains get separate copies of the mail. This is
intended to keep cleaner 'To' lines when sending mail to
archive-type lists.

This script now generates its own Message-Id header, so that
if more than one piece of mail is generated, they can easily
be detected by duplicate-suppression features of mailers.