Change rindex() to strrchr()
Default builds of uClibc (used for the Linux MFS) don't include
support for index() and rindex() (gone from the POSIX standard as of
POSIX.1-2008).  I've fixed the uClibc config to include them now, but
I need to build frisbee for an existing build of the Linux MFS with
the old uClibc.