Allow for more flexible setup of pxe_boot_path.
If nodes.pxe_boot_path is set to '/tftpboot/pxelinux/<something>', then
dhcpd_makeconf will set the (pxeboot) filename to /tftpboot/pxelinux.0
and symlink the node's config file (/tftpboot/pxelinux.cfg/<mac>) to
/tftpboot/pxelinux.cfg/<something>.

In other words, we can customize pxelinux to some small degree, using one
of some small number of pre-existing configurations. We were using pxelinux
before for plab-in-elab and we will also need it for loading WinPE for
configuring Windows7 images. For the latter we will set the pxe_boot_path
to /tftpboot/pxelinux/winpe.

Anyway, ideally we would allow the user to specify a pxelinux config file
through the NS file, but need to think about the implications of that some
more. Small steps...