Major fix release
Disabled link buttons on the slice canvas for managers without links. Switched delay node to be at the end of the draggable nodes on the slice window. Dragging a manager onto the canvas now adds a node or prompts the user to add one. Bolded the map markers.  General bug fixes various places.