Added more caching features
Scroll doesn't zoom map when mouse is over an info window. Slices in the map legend update with status and when selected. Added buttons to refresh slice, slices or everything. Fixed offline support. Added ability to use cache when logged in to use cached managers but live data for everything else.