Add -f option for reporting hashes.
This forces imagehash to hash pieces on aligned sector boundaries rather
than just doing it from the start of each chunk. Result now is that there
may be a partial hash piece at the beginning of each range and another
partial piece at the end. Ideally, we would avoid those partial pieces,
but we are only using this as a point of comparison right now.