A couple of fixes to make a FreeBSD 6.3 install work.
We don't officially support this anymore, but the problems were minor:

mapper: use HAVE_XERCES to determine whether rspec v2 output is supported.
The symptom is that swapin will fail due to passing the unsupported 'f'
option to assign, but it just isn't going to work anyway without an XML
parser installed.

rc.mkelab: fix a broken version check that would lead to inner-boss
setup failure on a pre-FreeBSD7 machine.