Fixes that allow you to add and delete links properly, with the manifest updated with new interfaces in the nodes section.
What you cannot do, is change the endpoint of an existing link. For
example, if link0 goes from A to B, you cannot submit an rspec that
changes link0 to go from A to C. What you *can* do is submit an rspec
that removes link0 and adds link1 from A to C. This was not working
before, for a number of reasons having to do with the manifest
generation, but those are all fixed now.