Add errorlog text field to aggregates and slivers so that non-node slivers can store/return error info to the user in sliverstatus. This include the top level aggregate for a slice.
Added an 'error' attribute to the top level sliverstatus return hash,
to return this error.

Put start/restart sliver into the background so that it runs async,
just like it does when invoked from the CreateSliver() path. This is
because start/restart can take an arbitrary length of time, and having
the RPC sit in hold for that long is not an ideal interface. Users can
get all the info they need from the sliverstatus call.

Change Start/Restart so that all of the error message we were printing
to STDERR for the mail log, also go back to the user in the toplevel
error for the aggregate.