Add a "gencabundle" script to generate ProtoGENI CA certificate bundles.
It only makes sense to run this at the clearinghouse.  It brings both
the local CA bundle and the bundle to be distributed to federates up
to date with respect to .../etc/genicacerts/*.pem and
.../etc/extracerts.bundle.

Any time the sources are changed, just run this script.  The local bundle is
ready immediately.  The federates will fetch the new version as they get
around to it.