New script to re-write history of git moves
This script is intended for use with 'git filter-branch' - it takes
a directory, finds all files that have been moved into that directory
and where they came from, and produces a script that moves (in the
filesystem sense, not the git sense) them from their old locations
to their new ones.

Then, you give this script to 'git filter-branch --tree-filter', and
it will re-write all history so that it looks like the files were
always in their new locations.

Written specifically for the move of the clientside stuff into its
own directory, and then off into its own repository.