Add vnode_id index to vinterfaces table.
Part I of "Lessons learned from a 100K node experiment". nfree would
run for hours trying to free 100K virtual nodes due to a slow query in
Node::ReleaseSharedBandwidth(). Now it takes 5 minutes.

There were other lessons as well, but they fall in the category of
"rearranging deck chairs" since we are going to have to massively overhaul
lots of infrastructure to support O(100000+) node experiments on a regular
basis.