Add support for sub-authorities.
Generate authority certificates for local sub-authorities (i.e. authorities
corresponding to a local project) on demand.

Map per-project URLs to the same XMLRPC server handled in the context of
the authority for the specified project.

Make the SA give out per-project credentials when it's asked for a
GetCredential in a sub-authority.