Ensure tbprepare runs after a reboot_prepare to capture the image.
If this Linux is using systemd as its init, make sure to "start" the
tbprepare service.  systemd will not honor the K89tbprepare rc*.d
symlinks *unless the tbprepare service has been started* -- and of
course it doesn't "start" via an S89tbprepare link.  So we make sure
systemd thinks it got started, and then the K89tbprepare links will
be run appropriately.