After DETER was recently upgraded to FreeBSD8.2 (with corresponding ports) we noticed that there was an incompatible change in p5-RPC-XML-0.65 (from 7.2) and -0.76 (from 8.2).
Formerly, RPC::XML::Parser() was OK with parsing only suitably
encoded arguments rather than entire MethodRequest or MethodResponse;
now it throws an error.

We've tested the following (2 + 1) line patches on FreeBSD7.2 and
know them to be backwards compatibile there, and highly suspect it will
be OK for 6.3, although we don't have the means of generating a 6.3
e-in-e anymore.