Add node/node_type attribute "no_clflush".
This is for FreeBSD to turn off use of the CLFLUSH instruction via the
hw.clflush_disable loader var. We have encountered some opteron machines
on which this causes problems.

The attribute can be set for the node_type (via the web page for editing
node types) or the node (via brute-force SQL hackary) to override whatever
FreeBSD would do by default. This attribute is passed via tmcd to the
diskloader MFS which will fixup the loader.conf file post-frisbee.