Use boss as the ssh proxy if inner ops is a VM.
Subtle: if your elabinelab uses a VM for ops, then there will be no
myops.eid.pid.emulab.net DNS alias created outside (since there is no actual
ops node in the topo) and thus the inner ops cannot be used as a proxy for
ssh'ing to the unroutable inner nodes. In that case we use the inner boss
instead.