More work on image permissions; allow specification of pid/osname in NS files. Tweak permission check in Geni CM to also allow this, although at this time only global images from any project are allowed. The virt_nodes table has been changed to accommodate pid/osname syntax:
	tb-set-node-os $nodeA somepid/someos

Note: we are really exporting permission to use images, not entries in
the os_info table (OSIDs) which is what the NS parser and protogeni CM
are using. But in fact, an image is both an image descriptor and an OS
descriptor linked together, so if you export an image or make it
global, you are implicitly doing the same for the OS descriptor. As
mentioned many times in the past, OSIDs suck.