Add -Z option to zero free space included in the image due to -F.
Previously, this "internal free space" was whatever garbage happened to
be on the disk we were imaging. By zeroing, we eliminate any leakage of
information from the source disk and also allow the image to compress just
a tad (1-4%) better.

Why isn't this the default? Eh...no good reason, other than that this whole
squish-out-small-free-ranges-to-allow-longer-writes optimization should be
done by the client and should not be encoded in the image.