A couple of fixes for the FreeBSD 8.x-based MFSes.
We apparently were not setting GID/EGID quite right in runstartup, it was
dropping a group. If that group happened to be "wheel", then any command
requiring "sudo" would fail. Now we set GID and EGID according to the letter
of the law.

If mount of a Linux FS fails, run e2fsck and try again. BSD cannot mount
a Linux FS if the ext3 journal requires recovery. e2fsck will handle that.